↩️ revert stale OCI MCP controller split
