+++ libs/backend/agents/runtime/mcp-executor/protocol/src/mcp-executor-protocol.validator.ts
@@ -1,13 +1,6 @@
-/**
- * Validates MCP server data before the executor stores it or returns it across its JSON boundary.
- * Keeping the schemas beside the protocol types lets a reader review the accepted wire form with
- * the durable representation that the executor returns.
- */
 import { z } from "zod";
 
 import type { JsonValue } from "@opencrane/util";
-
-import type { McpExecutorDiscoveredTool } from "./mcp-executor-protocol.types";
 
 /** Recursively accepts only values that can cross the executor's JSON boundary. */
 const _JsonValueSchema: z.ZodType<JsonValue> = z.lazy(function _JsonValue(): z.ZodType<JsonValue>
@@ -59,35 +52,16 @@ const _ContentBlockSchema = z.discriminatedUnion("type", [_TextContentSchema, _I
 
 const _ToolInputSchema = z.record(_JsonValueSchema).refine(function _ObjectSchema(value): boolean { return value["type"] === "object"; });
 
-/** Strict durable representation of one tool discovered from an MCP server. */
-const _DiscoveredToolSchema = z.object({
-	name: z.string().min(1).max(128),
-	description: z.string().max(4_096).nullable().optional(),
-	inputSchema: _ToolInputSchema,
-}).strict();
-
-/** Bounds a tools/list response before its names are checked for uniqueness. */
-const _DiscoveredToolsSchema = z.array(_DiscoveredToolSchema).max(256);
-
 /**
- * Parses the server-provided durable form of discovered MCP tools.
+ * Returns whether a live tool input is one JSON Schema object.
  *
- * The schema rejects unknown fields before the protocol parser checks duplicate names. The latter
- * remains a separate semantic rule because it applies across the complete returned list.
- *
- * Called by: __ParseMcpExecutorDiscoveredTools.
- * @param value - Untrusted `tools/list` result field.
- * @returns Strictly shaped tools without duplicate-name interpretation.
+ * {@link __ParseMcpExecutorToolsListResponse} uses this check before returning a discovered tool.
+ * @returns True when the value is a JSON object whose `type` field is `object`.
+ * @see https://modelcontextprotocol.io/specification/2026-07-28
  */
-export function _ParseMcpExecutorDiscoveredTools(value: unknown): readonly McpExecutorDiscoveredTool[] | null
+export function _IsMcpExecutorToolInputSchema(value: unknown): value is Readonly<Record<string, JsonValue>>
 {
-	const result = _DiscoveredToolsSchema.safeParse(value);
-	if (!result.success)
-		return null;
-	return result.data.map(function _Tool(tool): McpExecutorDiscoveredTool
-	{
-		return { name: tool.name, description: tool.description ?? null, inputSchema: tool.inputSchema };
-	});
+	return _ToolInputSchema.safeParse(value).success;
 }
 
 /**
