+++ libs/backend/server/iam/authorization/main/src/prisma-mcp-tool-invocation-participant.ts
@@ -2,18 +2,62 @@ import { type Prisma } from "@prisma/client";
 
 import type { JsonValue } from "@opencrane/util";
 
-import { _AppendMcpToolInvocationCompleted, _AppendMcpToolInvocationFailed, _EnterMcpToolInvocationRecovery, _MCP_AMBIGUOUS_FAILURE_CODE } from "./mcp-tool-invocation-lifecycle-events";
-import type { McpToolInvocationTransactionParticipant, McpToolInvocationTransactionParticipantFactory } from "./mcp-tool-invocation-participant.types";
 import { PrismaToolInvocationRepository } from "./prisma-tool-invocation-repository";
-import { ExternalActionClaimKinds, ToolInvocationStates } from "./tool-invocation-lifecycle.types";
-import { ToolInvocationCompletionOutcomes } from "./tool-invocation.types";
-import type { ToolInvocationClaim, ToolInvocationClaimResult, ToolInvocationCompletionResult, ToolInvocationLifecycleEventSink, ToolInvocationRecord, ToolInvocationRecoveryEventSink, ToolInvocationRunRecoveryAuthority, ToolResultDeliveryPayload } from "./tool-invocation.types";
+import { TOOL_INVOCATION_PREPARATION_POLICY, ExternalActionClaimKinds, ToolInvocationStates } from "./tool-invocation-lifecycle.types";
+import { ToolInvocationCompletionOutcomes, ToolInvocationEventTypes, ToolInvocationRunRecoveryEnterResults } from "./tool-invocation.types";
+import type { McpToolInvocationTransactionParticipant, McpToolInvocationTransactionParticipantFactory } from "./mcp-tool-invocation-participant.types";
+import type { ToolInvocationClaim, ToolInvocationClaimResult, ToolInvocationCompletionResult, ToolInvocationLifecycleEvent, ToolInvocationLifecycleEventSink, ToolInvocationRecord, ToolInvocationRecoveryEvent, ToolInvocationRecoveryEventSink, ToolInvocationRunRecoveryAuthority, ToolInvocationRunRecoveryEnterResult, ToolResultDeliveryPayload } from "./tool-invocation.types";
+
+/** Failure code stored when the companion cannot prove what the MCP server did. */
+const _AMBIGUOUS_FAILURE_CODE = "external_action_provider_outcome_ambiguous";
+
+/** Build the timeline event written after a checked MCP tool result succeeds. */
+function _CompletedEvent(invocation: ToolInvocationRecord): ToolInvocationLifecycleEvent
+{
+	return { runId: invocation.runId, attempt: invocation.attempt, eventType: ToolInvocationEventTypes.Completed, payload: { toolInvocationId: invocation.toolInvocationId } };
+}
+
+/** Build the timeline event written after a definite or uncertain MCP tool failure. */
+function _FailedEvent(invocation: ToolInvocationRecord, reason: string, retrying: boolean): ToolInvocationLifecycleEvent
+{
+	return { runId: invocation.runId, attempt: invocation.attempt, eventType: ToolInvocationEventTypes.Failed, payload: { toolInvocationId: invocation.toolInvocationId, toolRevisionId: invocation.toolRevisionId, reason, retryCount: invocation.preparationAttempt, retryLimit: TOOL_INVOCATION_PREPARATION_POLICY.attemptLimit, retrying } };
+}
+
+/** Append the timeline event or abort the transaction when the owning run refuses it. */
+async function _AppendLifecycleEvent(sink: ToolInvocationLifecycleEventSink, transaction: Prisma.TransactionClient, event: ToolInvocationLifecycleEvent): Promise<void>
+{
+	if (!await sink.appendInTransaction(transaction, event))
+		throw new Error("tool invocation transition requires its canonical lifecycle event");
+}
+
+/** Append the manual-recovery event or abort the transaction when the owning run refuses it. */
+async function _AppendRecoveryEvent(sink: ToolInvocationRecoveryEventSink, transaction: Prisma.TransactionClient, invocation: ToolInvocationRecord): Promise<void>
+{
+	const event: ToolInvocationRecoveryEvent = { runId: invocation.runId, expectedAttempt: invocation.attempt, toolInvocationId: invocation.toolInvocationId, preparationRetryCount: invocation.preparationAttempt, preparationRetryLimit: TOOL_INVOCATION_PREPARATION_POLICY.attemptLimit, providerOutcome: "unknown_after_dispatch" };
+	if (!await sink.appendInTransaction(transaction, event))
+		throw new Error("tool recovery state requires its canonical recovery event");
+}
+
+/** Move the run into recovery and write the matching event inside the same transaction. */
+async function _EnterRecoveryRequired(authority: ToolInvocationRunRecoveryAuthority, sink: ToolInvocationRecoveryEventSink, transaction: Prisma.TransactionClient, invocation: ToolInvocationRecord): Promise<void>
+{
+	const outcome: ToolInvocationRunRecoveryEnterResult = await authority.enterRecoveryRequiredInTransaction(transaction, { runId: invocation.runId, attempt: invocation.attempt });
+	if (outcome === ToolInvocationRunRecoveryEnterResults.Entered || outcome === ToolInvocationRunRecoveryEnterResults.AlreadyRecoveryRequired)
+	{
+		await _AppendRecoveryEvent(sink, transaction, invocation);
+		return;
+	}
+	if (outcome === ToolInvocationRunRecoveryEnterResults.Cancelling)
+		return;
+	throw new Error("tool recovery state conflicts with its owning run attempt");
+}
 
 /**
- * Applies authorization-owned MCP state changes inside the transaction supplied by the MCP runtime.
+ * Applies the authorization half of an MCP unit of work inside the transaction the MCP package owns.
  *
- * This unit of work owns ToolInvocation persistence but delegates matching timeline and recovery
- * writes through existing runs-owned ports. It never opens a transaction.
+ * This adapter is a unit-of-work participant rather than a repository because it coordinates the
+ * ToolInvocation row, result delivery, lifecycle event, and run recovery writers. It never opens a
+ * transaction; the caller must create a new instance for each transaction callback.
  */
 export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvocationTransactionParticipant
 {
@@ -28,7 +72,7 @@ export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvo
 	/** Transaction shared with the MCP runtime write. */
 	private readonly _transaction: Prisma.TransactionClient;
 
-	/** Binds every authorization writer to the transaction already opened by the MCP authority. */
+	/** Bind every authorization writer to the transaction already opened by the MCP authority. */
 	constructor(transaction: Prisma.TransactionClient, lifecycleEvents: ToolInvocationLifecycleEventSink, recoveryEvents: ToolInvocationRecoveryEventSink, runRecovery: ToolInvocationRunRecoveryAuthority)
 	{
 		this._transaction = transaction;
@@ -38,19 +82,19 @@ export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvo
 		this._runRecovery = runRecovery;
 	}
 
-	/** Returns the invocation row owned by the authorization package. */
+	/** Return the invocation row owned by the authorization package. */
 	async findById(invocationId: string): Promise<ToolInvocationRecord | null>
 	{
 		return this._repository.findById(invocationId);
 	}
 
-	/** Claims dispatch because the companion is about to call the uploaded MCP server. */
+	/** Claim dispatch, because the companion is about to call the uploaded MCP server. */
 	async claim(invocationId: string, now: Date, leaseMilliseconds: number): Promise<ToolInvocationClaimResult>
 	{
 		return this._repository.claim(invocationId, ExternalActionClaimKinds.Dispatch, now, leaseMilliseconds);
 	}
 
-	/** Saves a checked success and its timeline event without leaving the caller's transaction. */
+	/** Save a checked success and its event without leaving the caller's transaction. */
 	async completeSucceeded(claim: ToolInvocationClaim, result: JsonValue, now: Date): Promise<ToolInvocationCompletionResult>
 	{
 		const invocation = await this._repository.findById(claim.invocationId);
@@ -59,11 +103,11 @@ export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvo
 		const payload: ToolResultDeliveryPayload = { toolInvocationId: invocation.toolInvocationId, outcome: "succeeded", result };
 		const completed = await this._repository.complete(claim, payload, now);
 		if (completed.outcome === ToolInvocationCompletionOutcomes.Completed)
-			await _AppendMcpToolInvocationCompleted(this._lifecycleEvents, this._transaction, completed.invocation);
+			await _AppendLifecycleEvent(this._lifecycleEvents, this._transaction, _CompletedEvent(completed.invocation));
 		return completed;
 	}
 
-	/** Saves a definite failure and its timeline event without leaving the caller's transaction. */
+	/** Save a definite failure and its event without leaving the caller's transaction. */
 	async completeFailed(claim: ToolInvocationClaim, failureCode: string, now: Date): Promise<ToolInvocationCompletionResult>
 	{
 		const invocation = await this._repository.findById(claim.invocationId);
@@ -72,11 +116,11 @@ export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvo
 		const payload: ToolResultDeliveryPayload = { toolInvocationId: invocation.toolInvocationId, outcome: "failed", failureCode };
 		const completed = await this._repository.complete(claim, payload, now);
 		if (completed.outcome === ToolInvocationCompletionOutcomes.Completed)
-			await _AppendMcpToolInvocationFailed(this._lifecycleEvents, this._transaction, completed.invocation, completed.invocation.failureCode ?? "external_action_failed", false);
+			await _AppendLifecycleEvent(this._lifecycleEvents, this._transaction, _FailedEvent(completed.invocation, completed.invocation.failureCode ?? "external_action_failed", false));
 		return completed;
 	}
 
-	/** Saves an uncertain outcome before it enters the existing run-owned recovery flow. */
+	/** Save an uncertain outcome and enter the existing run recovery flow. */
 	async completeAmbiguous(claim: ToolInvocationClaim, now: Date): Promise<ToolInvocationRecord | null>
 	{
 		const transition = await this._repository.completeAmbiguous(claim, now);
@@ -84,9 +128,9 @@ export class PrismaMcpToolInvocationParticipantUnitOfWork implements McpToolInvo
 			return transition.invocation;
 		const invocation = transition.invocation;
 		const retrying = invocation.state === ToolInvocationStates.Ready || invocation.state === ToolInvocationStates.Reconciling;
-		await _AppendMcpToolInvocationFailed(this._lifecycleEvents, this._transaction, invocation, _MCP_AMBIGUOUS_FAILURE_CODE, retrying);
+		await _AppendLifecycleEvent(this._lifecycleEvents, this._transaction, _FailedEvent(invocation, _AMBIGUOUS_FAILURE_CODE, retrying));
 		if (invocation.state === ToolInvocationStates.RecoveryRequired)
-			await _EnterMcpToolInvocationRecovery(this._runRecovery, this._recoveryEvents, this._transaction, invocation);
+			await _EnterRecoveryRequired(this._runRecovery, this._recoveryEvents, this._transaction, invocation);
 		return invocation;
 	}
 }
